Fixed status codes

+++ routes/orders.js
@@ -24,13 +24,13 @@ function getOrders() {
 
 router.get('/', (req, res) => {
     if (req.session.loggedin) {
-        logger.log('Getting orders from database...');
+        logger.info('Getting orders from database...');
 
         res.json(getOrders());
     }
     else {
         logger.error('User is not logged in!');
-        res.json({
+        res.status(401).json({
             error: 'You need to be logged in to access orders.'
         });
     }
@@ -58,7 +58,7 @@ router.get('/:id', (req, res) => {
     }
     else {
         logger.error('User is not logged in!');
-        res.json({
+        res.status(401).json({
             error: 'You need to be logged in to access orders.'
         });
     }
@@ -69,84 +69,100 @@ router.get('/:id', (req, res) => {
 //                                         POST REQUESTS
 //--------------------------------------------------------------------------------------
 router.post('/create', (req, res) => {
-    logger.log('Creating order ' + req.body.orderNumber);
+    if (req.session.loggedin) {
+        logger.log('Creating order ' + req.body.orderNumber);
 
-    const id = req.body.orderId;
-    const productId = req.body.productId;
-    const userId = req.body.userId;
-    const status = req.body.status ? req.body.status : 'busy';
-    const quantity = req.body.quantity;
-    const price = req.body.price;
-
-
-    if (!id || !productId || ! userId || !status || !quantity || !price) {
-        res.redirect('/?status=error');
-        logger.error('Failed to instert new order: some fields where empty.');
-        return;
-    }
-
-    const queryString = 'INSERT INTO orders (id, productId, userId, status, quantity, price) VALUES (?, ?, ?, ?, ?, ?)';
-    SQLConnection().query(
-        queryString,
-        [   
-            id,
-            productId,
-            userId,
-            status,
-            quantity,
-            price
-        ],
-        (err, result, fields) => {
-            if (err) {
-                logger.error('Failed to insert new order: ' + err);
-                res.status(500).redirect('/?status=error');
-                return;
-            } 
-
-            logger.success('Inserted new order with id: ' + result.insertId);
-            res.status(201).redirect('/?status=success');
+        const id = req.body.orderId;
+        const productId = req.body.productId;
+        const userId = req.body.userId;
+        const status = req.body.status ? req.body.status : 'busy';
+        const quantity = req.body.quantity;
+        const price = req.body.price;
+    
+    
+        if (!id || !productId || ! userId || !status || !quantity || !price) {
+            res.status(500).redirect('/?status=error');
+            logger.error('Failed to instert new order: some fields where empty.');
+            return;
         }
-    );
+    
+        const queryString = 'INSERT INTO orders (id, productId, userId, status, quantity, price) VALUES (?, ?, ?, ?, ?, ?)';
+        SQLConnection().query(
+            queryString,
+            [   
+                id,
+                productId,
+                userId,
+                status,
+                quantity,
+                price
+            ],
+            (err, result, fields) => {
+                if (err) {
+                    logger.error('Failed to insert new order: ' + err);
+                    res.status(500).redirect('/?status=error');
+                    return;
+                } 
+    
+                logger.success('Inserted new order with id: ' + result.insertId);
+                res.status(201).redirect('/?status=success');
+            }
+        );
+    }
+    else {
+        logger.error('User is not logged in!');
+        res.status(401).json({
+            error: 'You need to be logged in to access orders.'
+        });
+    }
 })
 
 //--------------------------------------------------------------------------------------
 //                                         UPDATE REQUESTS
 //--------------------------------------------------------------------------------------
 router.patch('/:id', (req, res) => {
-    logger.log('Updating order with id: ' + req.params.id);
+    if (req.session.loggedin) {
+        logger.log('Updating order with id: ' + req.params.id);
 
-    const id = req.body.orderId;
-    const status = req.body.status ? req.body.status : 'busy';
-    const quantity = req.body.quantity;
-    const price = req.body.price;
-
-
-    if (!id || !status || !quantity || !price) {
-        res.redirect('/?status=error');
-        logger.error('Failed to instert new order: some fields where empty.');
-        return;
-    }
-
-    const queryString = 'UPDATE orders status = ?, quantity = ?, price = ? WHERE id = ?';
-    SQLConnection().query(
-      queryString,
-      [
-        status,
-        quantity,
-        price,
-        id
-      ],
-      (err, result, fields) => {
-          if (err) {
-              logger.error('Failed to update order with id: ' + req.params.id)
-              res.status(500).redirect('/?status=error');
-              return;
+        const id = req.body.orderId;
+        const status = req.body.status ? req.body.status : 'busy';
+        const quantity = req.body.quantity;
+        const price = req.body.price;
+    
+    
+        if (!id || !status || !quantity || !price) {
+            res.status(500).redirect('/?status=error');
+            logger.error('Failed to instert new order: some fields where empty.');
+            return;
+        }
+    
+        const queryString = 'UPDATE orders status = ?, quantity = ?, price = ? WHERE id = ?';
+        SQLConnection().query(
+          queryString,
+          [
+            status,
+            quantity,
+            price,
+            id
+          ],
+          (err, result, fields) => {
+              if (err) {
+                  logger.error('Failed to update order with id: ' + req.params.id)
+                  res.status(500).redirect('/?status=error');
+                  return;
+              }
+    
+              logger.success('Updated order with id: ' + req.params.id);
+              res.status(200).redirect('/?status=success');
           }
-
-          logger.success('Updated order with id: ' + req.params.id);
-          res.status(200).redirect('/?status=success');
-      }
-    );
+        );
+    }
+    else {
+        logger.error('User is not logged in!');
+        res.status(401).json({
+            error: 'You need to be logged in to access orders.'
+        });
+    }
 });
 
 
@@ -154,19 +170,27 @@ router.patch('/:id', (req, res) => {
 //                                         DELETE REQUESTS
 //--------------------------------------------------------------------------------------
 router.delete('/:id', (req, res) => {
-    logger.log('Deleting order with id: ' + req.params.id);
+    if (req.session.loggedin) {
+        logger.log('Deleting order with id: ' + req.params.id);
 
-    const queryString = 'DELETE FROM orders WHERE id = ?';
-    SQLConnection().query(queryString, [req.params.id], (err, reslut, fields) => {
-        if (err) {
-            logger.error('Failed to delete order with id: ' + req.params.id);
-            res.status(500).redirect('/?status=error');
-            return;
-        }
-
-        logger.success('Deleted order with id: ' + req.params.id);
-        res.status(200).redirect('/?status=success');
-    });
+        const queryString = 'DELETE FROM orders WHERE id = ?';
+        SQLConnection().query(queryString, [req.params.id], (err, reslut, fields) => {
+            if (err) {
+                logger.error('Failed to delete order with id: ' + req.params.id);
+                res.status(500).redirect('/?status=error');
+                return;
+            }
+    
+            logger.success('Deleted order with id: ' + req.params.id);
+            res.status(200).redirect('/?status=success');
+        });
+    }
+    else {
+        logger.error('User is not logged in!');
+        res.status(401).json({
+            error: 'You need to be logged in to access orders.'
+        });
+    }
 });
 
 module.exports = router;
